Added basic login.
Added login function in userscontroller which validates the email/pw
combination and returns a random user token if it is valid.
Added a before_action that applies to all requests, except login, new
and create user requests. This action validates the provided token and
checks it's expiration date.
Added /t/:token to all routes; removed resource users and added it's
routes manually.
Added pw, token and expirationDate to the database schema.